refactor: wire in new key serde
Switch the code from using the old `SerdeFactory` for key serdes to using the same mechanism as value serdes, which brings with it support for types other than `String` and formats other than `Kafka`.  A follow up change will test and enable these additional features.

The change also switches the key in the code from a `String` to a `Struct`.