refactor: decouple ast/expression visitors (#3151)
refactor: decouple ast/expression visitors

This patch is the last step in completely decoupling the ast and expression
visitors:

Removes all the visitXExpression calls from AstVisitor

Gets rid of DefaultAstVisitor - none of its default behaviour is used

Remove all the visitXExpression calls from StatementRewriter. Instead,
StatementRewriter accepts a BiFunction<Expression, C, Expression> which it
uses to rewrite expressions.

Simplify the ExpressionTreeRewriter. Rather than having a sister interface
which is nearly identical to a visitor, this patch changes
ExpressionTreeRewriter to just accept a plugin which it will use to optionally
rewrite each expression in the tree (based on whether the returned optional
is empty or not). The plugin is always implemented as a visitor, but that's
not really important to understanding ExpressionTreeRewriter.

Implement StatementRewriteForStruct using a plugin for ExpressionTreeRewriter.