test: disable ZK admin server
This change disables the Zookeeper Admin Server that would otherwise be started by `ZookeeperEmbedded`.
The admin server defaults to using port 8080, which causes a port clash if two tests/instances are running on the same machine.
On a port clash the admin server fails to start, which causes Zookeeper to fail too.