refactor: remove pointless internal column name to index map
The `internalNameToIndexMap` map in `AggregateNode`'s `InternalSchema` class only ever contains mappings such as `{"KSQL_INTERNAL_COL_0" -> 0, "KSQL_INTERNAL_COL_1" -> 1}`.  This adds complexity to the code therefore makes it harder to understand what is going on.  This change removes this mapping and instead gets the index from the number at the end of the column name.