refactor: build aggregate functions only once
The change means the AggregateNode only builds the aggregate functions once.

Before it would build them first to get the return type needed to build the schema, then again to pass to the `aggregate` function.

This change reverses the order of these two steps, so now the schema can be built using the same aggregate function instances that will be later passed to `aggregate`.