refactor: cleanup interface b/w planner and structured (#3228)
refactor: cleanup interface b/w planner and structured

This patch aims to clean up the interface between the logical plan
nodes an structured, as a precursor to having SchemaKStream/Table
build the underlying execution plan. This entails the following:

Firstly, we move the code in DataSourceNode that deals with creating
the initial KStream and adding the key and timestamp columns to the
value to StreamSource.build. This is invoked by a new static factory
method for SchemaKStream, which creates the source SchemaKStream for
a given data source. Additionally, the SchemaKStream constructor has
been made package-private, so the only way to create these now is
from a source (fromSource), or via one of the SchemaKStream transforms.

Next, the code to convert a stream into a table has been moved to a
method in SchemaKStream called toTable, which applies the grouping
andd aggregation that was previously applied directly from
DataSourceNode.

Finally, this patch cleans up the other usage of the SchemaKStream
constructor (in KsqlStructuredDataOutputNode), and covnerts it to a
call to a new transform called `sink`.