feat: wire in the KS config needed for point queries (MINOR) (#3251)
* feat: wire in the KS config needed for point queries (MINOR)

Point queries need to be able to determine with KSQL node stores the data for a specific key in a state store. Kafka Streams supports this via its interactive queries functionality. For this to work, the KS `application.server` config must be set.

This change sees the `application.server` being automatically set by the rest application to the FIRST listener supplied in the standard `listener` config.

The `listener` can use automatic port allocation, e.g. if set to `http://localhost:0` then it will auto-allocate a port number. This means the port is not known until the server has been started, which requires the `KsqlConfig` to be injected into certain types once the server is started.