test: add RestQueryTranslationTest
This is along the same lines as `QueryTranslationTest`, but comes in through the front door of the REST Api.

It's currently much more limited in its functionality, but can extended as needed to be it more inline with QTT functionality.

QTT should still do the brunt of the work, with RQTT only being used where QTT will no suffice, e.g. `INSERT VALUE` statement test, or the upcoming static query work.