refactor: move column name validation out of the parser (MINOR)
Previously, the parser would throw if C* statements had columns with reserved words or if the key field was not `ROWKEY STRING`.

These checks should not be in the parser, but in the engine as they are requirements of the engine.

This also allows the SchemaParser to handle parsing schemas with multiple key fields, which is needed for the static query work.