refactor: move the statement rewrites to the engine
We have to statement re-writers. These currently reside in the parser, but more correctly belong in the engine.

First, we have `StatementRewriteForStruct`, which searches out and replaces any `DereferencedExpression` with a special UDF for accessing fields in structs.
Previously, this re-write happened in the parser. This PR sees this logic move into the engine. After all, it's an implementation of the engine that we handle struct field access via a UDF.

Second, we have `StatementRewriteForRowtime` which looks to allow comparisons between rowtime and strings, given we don't have a datatime type yet.
This is already only used in the engine. So moving the type itself to the engine seems sensible.

While moving, the API to these types was cleaned up. The `if(statement needs rewrite)` check has been moved inside the re-write class and the class has been refactored to make it more suitable for dependency injection.