fix(3356): struct rewritter missed EXPLAIN (#3398)
* fix(3356): struct rewritter missed EXPLAIN

KSQL uses a statement rewritter to convert STRUCT field dereferences into `FETCH_FIELD_FROM_STRUCT` function calls.  The rewrite is done for `Query`, `C*AS` and `INSERT INTO` statements, but was missing `EXPLAIN`.

It is valid to execute `EXPLAIN` on a query string, e.g.  `EXPLAIN SELECT address->street FROM Y;`.  It is therefore important that the query within the `EXPLAIN` statement has the same rewrites applied to it as would be applied should the statement be executed directly.