refactor: switch properties to composition over inheritance (#3406)
* refactor: switch properties to composition over inheritance

Why? Mainly because the code does not care about these two types having a common base class. `WithClauseDef` is just a helper. So composition is preferable to inheritance.

Also, it reduces the temptation to place inappropriate code in `WithClauseDef` which is intended as a generic helper, and should not know about specific properties.