feat(static): unordered table elements and meta columns serialization (#3428)
* feat(static): unordered table elements and meta columns serialization

A query might contain key and value columns in any order. Hence the `TableElements` class has been changed to:

 - not fail if key columns come after value columns
 - not change the order of columns

When serializing a logical schema across the wire, (which is currently only used for static queries), we should't yet serialize the ROWTIME meta column. This is mainly because static queries don't support it, and our syntax doesn't support meta columns, i.e. you can't have a `CREATE TABLE X (ROWTIME BIGINT META, ...`. So if we did include it in the serialized form, we wouldn't be able to parse it.