refactor: simplify the logic in EngineExecutor and QueryEngine for executing queries and ddl (#3448)
Previously the logic was somewhat convoluted with the call to create a logical plan occurring even if the statement was DDL (and with the redundant creation of a QueryEngine instance in this case).

I've simplified the logic so logical plan nodes are only created if the statement is a query or query container. Moreover, QueryEngine now is only concerned with queries and the instanceof checking has been moved out of QueryEngine into the EngineExecutor and simplified. The resulting logic is much clearer and easier to follow.