fix: fix test for KIP-307 final PR (#3402)
The fifth and final PR for KIP-307 is getting merged today. This PR adds the following overload KGoupedTable.aggregate(Initializer, Aggregator, Aggregator, Named).

This causes an error with the SchemaKGroupedTableTest when executing mockKGroupedTable.aggregate(any(), any(), any(), any())) method. The compiler can't disambiguate the 4th argument, as it's either a Materialized or a Named object. This PR fixes the test.