fix: improve print topic error message when topic does not exist
The message incorrectly stated that KSQL treats unquoted topic names as uppercase. This is not the case.

Also, the suggested alternative was just the opposite case to the supplied, which is fairly arbitrary, e.g. if the user issues `print PageView` and the topic does not exist the error message will suggest `pAGEvIEW`, even though no such topic exists.

With this change the error message will include any other topics that exist in the cluster with the same name,  but different case.