refactor: split out ddl creation into factories (#3447)
This patch is a follow up to a previous commit which moved the logic
for initializing DDL commands out of the command pojos and into
CommandFactories. In this patch, we move this logic out into separate
factory classes for the different ddl types. CommandFactories is now
just a way to route arbitrary ddl statements to the right factory.