refactor: get the physical plan (almost) out of the engine (#3465)
This patch refactors the engine and physical plan builder so that
we can return the query plan from the engine:

First, we define the pojos that define the plan:
   - PhysicalPlan: this contains all the information needed to
     reconstruct the kafka streams app, and is the new return type
     for PhysicalPlanBuilder.
   - QueryPlan: this contains the PhysicalPlan, and other information
     needed about a query, such as the set of data sources read, and
     the data source written into.
   - KsqlPlan: this contains an optional QueryPlan, and an optional
     DdlCommand. A given DDL/DML statement can contain one or both of
     these.

PhysicalPlanBuilder has been simplified to just return a PhysicalPlan,
but not actually build the plan into a Kafka Streams app.

EngineExecutor has been rearranged to support building/executing KsqlPlan,
in addition to ConfiguredStatement. It's new interface has:
   - plan(ConfiguredStatement<?> statement): given a ddl or dml, build a
     KsqlPlan instance.
   - execute(KsqlPlan plan): given a plan, execute the plan (apply any
     ddls, and create a PersistentQueryMetadata)
   - execute(ConfiguredStatement<?> statement): fully execute a given
     statement.
   - executeQuery(ConfiguredStatement<Query> statement): fully execute
     a transient query.

Under the hood, EngineExecutore uses a new class called QueryExecutor to
execute query plans from a physical plan. QueryExecutor is responsible for
producing the final streams app and wrapping it in an instance of
QueryMetadata. QueryExecutor is basically doing most of what
PhysicalPlanBuilder used to do.