feat(static): unique id per static query
Add a unique(ish) id for each static query, based on the current time.

Any entries in the processing log will be tagged with this query id, which will aid debugging.

There's obviously still potential for two queries to end up with the same id, but this is unlikely and will not cause any issues if it does happen, except for maybe making it a little harder to debug any issues.