feat: build materializations from the physical plan
This patch changes up how we build materializations to use the
physical plan instead of building from the logical plan. A
materialization is the data in the state store with some additional
transformations applied. These transformations should be the same
as those applied by the steps in the physical plan following the
aggregation step. This patch builds up the transformation as we
build the streams app from the physical plan.

Transformations are described in MaterializationInfo, and
accumulated by a Builder class. A transformation allows mapping
a key/row pair to an optional row. The row is optional to support
filters. We currently support 3 types of transformations:
   - aggregate result map: mapping the results of complex aggregates
     (e.g. avg)
   - projection
   - filter

Finally, the materialization code has been moved to the ksql-streams
module so that it can be used by the streams app builder.