refactor: Remove KsqlAggregateFunction.getInstance (#3483)
Previously, the way we obtained and initialised instances of KSqlAggregateFunction was over complex: We would obtain an instance with functionRegistry.getAggregate, and then call getInstance on the KsqlAggregateFunction to obtain a second instance of the class. The first instance was then garbage collected.

This was inefficient and led to unnecessarily complex code. It also meant the code generation of UDAFs was more complex than it needed to be.

It seems that the getInstance method was introduced with the TopkKudaf. This was the first to take an extra initialisation parameter (the number k in this case), and this value needs to be known by the udaf implementation. The old functionRegistry.getAggregateFunction API didn't allow initialisation parameters to be passed in, so the getInstance method was introduced to allow a new instance with the new params to be created from the old instance.

This refactoring removes KSqlAggregateFunction.getInstance and allows initialisation parameters to be passed to functionRegistry.getAggregate (which has been renamed to functionRegistry.getAggregateFunction as that is what it does). This means we don't need an extra getInstance method to create an instance with the right initialisation parameters. The udaf generation code is now simpler as we don't have to generate a getInstance method.