fix: changes required for compatibility with KIP-479 (#3466)
This PR introduces the StreamJoined object from KIP-479. Specifically this PR

Removes use of the now deprecated KStream#join(KStream,....Joined) methods in favor of KStream#join(KStream,....StreamJoined)
Keeps the name of the stores so none of the expected topologies needed changing. This was a judgment call on my part, and the KSQL team will need to let me know their preferences.