Combine rowkey rowtime steps in KS topology (#3613)
* test: extend test framework to allow bounds on KSQL version

This commit allows a new `versions` element to be added to test cases that controls the `min` and `max` versions of KSQL the test case is valid for.

This allows us to make known backwards-compatible changes to KSQL that would previously have caused an existing test to fail, e.g. changing the name of the internal topics.  With this commit we can mark the old test as only valid up to the last version and create a new version of the test for, with a min version, to test versions going forward

e.g.

```json
[{
      "name": "stream stream left join",
      "versions": {
        "max": "5.3"
      },
      "statements": [
        ...
      ],
      "inputs": [
        ...
      ],
      "outputs": [
        {"topic": "_confluent-ksql-some.ksql.service.idquery_CSAS_LEFT_OUTER_JOIN_0-KSTREAM-JOINTHIS-0000000008-store-changelog", "value": {"ROWTIME": 0, "ROWKEY": "0", "ID": 0, "NAME": "zero", "VALUE": 0}, "timestamp": 0},
        {"topic": "LEFT_OUTER_JOIN", "key": 0, "value": {"T_ID": 0, "NAME": "zero", "VALUE": 0, "F1": null, "F2": null}, "timestamp": 0}
      ]
    },
    {
      "name": "stream stream left join post v5.4",
      "versions": {
        "min": "5.4"
      },
      "statements": [
        ... same statements
      ],
      "inputs": [
        ... same inputs
      ],
      "outputs": [
        {"topic": "different-internal-topic-store-changelog", "value": {"ROWTIME": 0, "ROWKEY": "0", "ID": 0, "NAME": "zero", "VALUE": 0}, "timestamp": 0},
        {"topic": "LEFT_OUTER_JOIN", "key": 0, "value": {"T_ID": 0, "NAME": "zero", "VALUE": 0, "F1": null, "F2": null}, "timestamp": 0}
      ]
    }]
```

* refactor: combine steps for adding ROWTIME and ROWKEY

Previously, we had two steps on each source in the topology:

1. a `mapValues` call to add ROWKEY to the value
2. a `transformValues` call to add ROWTIME to the value

This commit combines them into a single `transformValues` call that does both.

For backwards compatibility, a no-op `mapValues` call is added for existing queries.
This is handled via the new CompatibilityBreakingConfigDef `ksql.query.inject.legacy.map.values.node`