refactor: use the function call expression in flat map step
This patch changes the flat map execution plan step to contain a list of
function calls, rather than a list of appliers. This is preferable because
FunctionCall is part of the model that will be serialized to the command/config
topic. This is also consistent with other steps that execute functions (projection,
aggregation).