feat: plan serialization jackson annotations
This patch adds jackson annotations needed to serialize execution plans:
    - query context / query ID
    - DDLs
    - execution steps
    - name types

It also adds annotations for a few types that the above types depend on:
    - timestamp extractors
    - format pojos needed for serdes
These will be cleaned up in a later change.

Finally, this patch includes a tool for generating a JsonSchema spec from the
annotated types, a first cut of this schema, and a test for checking that the
schema isn't changed.