refactor: decouple metastore from ast builder
This is a prep PR for expression and schema serialization/deserialization using
KSQL's grammar. The goal here is to get the meta store (mostly) out of the parser,
so that we can use it to convert from antlr's AST to KSQL's when parsing expressions
and schemas. AstBuilder still uses the type registry, however schema and expression
parsers can pass an empty type registry.

Before this change AstBuilder was using the metastore to do some validation, and to
fill in source qualifiers and column aliases. This has been moved into a statement
rewriter called AstSanitizer, which takes ASTs generated from externally-provided
statements and applies this additional validation and fills in these missing bits
expected by the engine.

Furthermore, this verification/rewrite is done by the engine now instead of the
parser. Currently this is happening in the engine's prepare call.