feat: drop legacy key field functionality
BREAKING CHANGE

This drops functionality that maintained a legacy version of the key field.
This legacy version was often wrong and resulted in unnecessary repartition steps being performed.
The functionality was previously required to ensure existing legacy queries continued to have the unnecessary repartitions to ensure no data loss.

With a major version bump this can now be dropped.