chore: remove GenericRowValueTypeEnforcer
This change sees `GenericRowValueTypeEnforcer` removed from the code base.

It's being removed as the functionality it provides is:

1. Not needed in the production code path as values are always of the right type.
1. Only needed for a few tests to pass, (which have been fixed up).
1. It supports coercing between types which are outside of KSQLs implicit casting rules, e.g. coerce String -> BIGINT.

We should support implicitly casting parameters to functions to wider types,
as defined by KSQLs implicit casting rules. However, this is not currently the
case. See https://github.com/confluentinc/ksql/issues/3791.