test: correct docs and extend QTT to allow it to run only against latest (#3749)
The docs around setting which topology versions to run when running the QTT test were wrong. The syntax is `-Dtopology.versions="5_0,5_3_0"`, not `-Dtopology.versions="5.0,5.3.0"`, i.e. the supplied versions must match the name of the folders under the `expected-topologies` directory.

Plus, there was no way that I could find to just run the latest version of the tests, so I've added the ability to set `-Dtopology.versions=latest-only`.