chore: rename `static` or `point` queries to `pull`
In both code and error messages.

Also, reorder some error messages to make them play better with the 127 character limit on closing a websocket connection, i.e. move the most important bit of a pull query error to the start of the error message so that it shows up in C3