refactor: ensure pull schema has no meta columns (#3907)
* refactor: ensure pull schema has no meta columns

Second part of the fix for https://github.com/confluentinc/ksql/issues/3859

Pull queries don't return ROWTIME, as this would be derived from the triggering source event in a push query, (i.e. ROWTIME comes from the timestamp of the event that triggers the output). Pull queries don't have a triggering source event, so don't have a ROWTIME. Hence pull query schemas should not have ROWTIME.

This commit ensures pull query schemas don't have ROWTIME and remove the temporary fix put in on the pull query websocket code to remove the meta columns

The commit adds a `withoutMetaColumns` to the `LogicalSchema`. Likewise, this won't remain for long, but for now this makes sense. Importantly, `withoutMetaColumns` maintains the order of columns within the schema.