refactor: bake known code for pull queries when building agg table (#3910)
* refactor: bake know code for pull queries when building agg table

This commit looks to bake the java code we know we'll need to to evaluate pull queries when we build the aggregate table itself.

To do this, I needed to change `SqlPredicate` and `SelectValueMapper` to accept the `ProcessingLogger` at a later point.

`MaterializationInfo` now holds the _actual_ aggregator, select mapper and having predicate, rather than just the info needed to create them. This simplifies the code and makes it much harder for these to get out of sink with the source table.