fix: dont create udtf sink w/ qualified rowkey
Fixes an issue where the sink stream (in the meta-store) for a query
w/ a UDTF has a rowkey column qualified with the original source
stream. Data sources in the meta-store should never have column
qualifiers. They certainly shouldn't have a mix of qualified and unqualified
columns.

This happens because the result of the FlatMap logical node has a
schema with both aliased and not-aliased columns. When creating
the final key column, the logical planner would decide whether or not
to strip off the alias from the key based on isAlias, which checks just
the first column for an alias. Instead, we just always remove the alias
from the key column when creating the final project schema.