chore: switch `TERMINATE ALL` to distribute N `TERMINATE queryId` stmts
As per discussion https://github.com/confluentinc/ksql/pull/3944#discussion_r349709433

This PR switching the `TERMINATE ALL` functionality to write a `TERMINATE <querId>` command to the command topic for each active query, rather than a single `TERMINATE ALL` command.  The reasoning behind this is that it we _want_ the command topic to be key compacted, and such a design works better with this plan.