fix: unify behavior for PARTITION BY and GROUP BY
BREAKING CHANGE: this change makes it so that PARTITION BY statements
use the _source_ schema, not the value/projection schema, when selecting
the value to partition by. This is consistent with GROUP BY, and
standard SQL for GROUP by. Any statement that previously used PARTITION
BY may need to be reworked.
BREAKING CHANGE: when querying with EMIT CHANGES and PARTITION BY, the
PARTITION BY clause should now come before EMIT CHANGES.