feat: remove unnecessary changelog for topics (#3987)
* refactor: remove unnecessary stream to table conversion

BREAKING CHANGE: KSQL will now, by default, not create duplicate changelog for table sources.

fixes: https://github.com/confluentinc/ksql/issues/3621

Now that Kafka Steams has a `KTable.transformValues` we no longer need to create a table by first creating a stream, then doing a select/groupby/aggregate on it. Instead, we can just use `StreamBuilder.table`.

This change makes the switch, removing the `StreamToTable` types and calls and replaces them with either `TableSource` or `WindowedTableSource`, copying the existing pattern for `StreamSource` and `WindowedStreamSource`.

It also reinstates a change in `KsqlConfig` that ensures topology optimisations are on by default. This was the case for 5.4.x, but was inadvertently turned off.

With the optimisation config turned on, and the new builder step used, KSQL no longer creates a changelog topic to back the tables state store. This is not needed as the source topic is itself the changelog.  The change includes new tests in `table.json` to confirm the change log topic is not created by default and is created if the user turns off optimisations.

This change also removes the line in the `TestExecutor` that explicitly sets topology optimisations to `all`. The test should _not_ of being doing tis. This may been why the bug turning off optimisations was not detected.

* chore: remove the reduce step from the expected schemas

The state store backing a table no longer requires explicit instantiation, so no value serde is being created, and its the creating of the value serde that results in a line in the expected schemas.

To put this another way, the state store now uses the same serde as the source topic.

* chore: rewrite expected topologies for tables

The expected topology for tables has changed and needs updating.

* chore: remove reduce step from expected schemas for joins

missed this previously

* chore: fix tests

* chore: post merge fix up

* chore: update JSON schema for command topic

* chore: fix test

* chore: fix test post merge

* chore: fix test post merge