BREAKING CHANGE: fix: dont update rowkey when executing StreamSelectKey (#4024)
* fix: dont update rowkey when executing StreamSelectKey

This patch changes execution of the StreamSelectKey step to no longer
support updating the ROWKEY column in the value, as it doesn't make
sense to do this from either place the stream is being rekeyed. Currently,
rekeys are done to rekey a stream for a join, or to rekey a stream to
satisfy a PARTITION BY clause. In either case, it doesn't make sense to
update the ROWKEY column in the value. That ROWKEY should contain the value
of the key in the data source.