refactor: logical plan nodes should not copy key & meta into value
The logical plan nodes should neither care, nor know, about the need of the implementation, i.e. our code around kafka-streams, needing to duplicate the key and meta columns into the value schema.  This PR changes the logical steps so that they deal purely with the logical schema.

Changes to the `Analyzer` mean:

1. Non-join persistent queries now reject statements where their is an unaliased `ROWKEY` or `ROWTIME` in the project, as this is a reserved column name.  Previously, they were silently dropped.
2. Non-join persistent queries now do not include `ROWTIME` and `ROWKEY` when expanding `SELECT *`.

This means the Analyzer is not letting through invalid select expressions, which gives better user feedback, and allows us to simplify the code, (e.g. see the removal of `getNoneMetaOrKeySelectAliases` in the `Analyzer`).

Also in this commit:

1. `LogicalSchema.isAliased` has been dropped as it's implementation was broken. We currently build schemas with a mixture of with and without source columns.  I'll address this in a follow up Pr.