fix: remove mapValues that excluded ROWTIME and ROWKEY columns (#4066)
* fix: remove mapValues that excluded ROWTIME and ROWKEY columns

With the changes introduced in #4052 it is no longer possible for the output schema of persistent queries to have reserved column names, i.e. `ROWKEY` or `ROWTIME` and hence the code to exclude these, both from the schema and from the output, is no longer needed and can be removed.

`ROWKEY` and `ROWTIME` are still copied into the value schema to ensure they are easily available. However, the final `ProjectNode` will never include them in the output as they are not allowed.