feat: support writing plans to command topic
This patch adds support for writing the physical plan to the command
topic in place of the original KSQL statements.

Part of this change is a small refactor to our rest-app to engine communication
module (ksql.rest.server.computation):
- CommandQueue.enqueue now accepts CommandId and Command as parameters rather
  than computing them.
- DistributingExecutor no longer takes a RequestValidator. Instead, it accepts
  a factory for creating validated commands. A validated command is one that has
  been validated to execute successfully against an engine snapshot.
  RequestValidator also uses an instance of this factory to validate statements
  at the api endpoint layer.

For now, writing of query plans is enabled by a config (ksql.execution.plan.enable)
which defaults to false. We'll drop this config once all the plan schema changes
are merged.